Added svn ignore tags.

+++ tags/leocad-0.76/win/Print.cpp
@@ -196,7 +196,7 @@ static void PrintCatalogThread (CWnd* pParent, CFrameWnd* pMainFrame)
 	
 	// Setting up the current OpenGL context
 	pfnwglMakeCurrent(pMemDC->GetSafeHdc(), hmemrc);
-	double aspect = (float)w/(float)h;
+	float aspect = (float)w/(float)h;
 	glMatrixMode(GL_MODELVIEW);
 	glViewport(0, 0, w, h);
 
@@ -299,7 +299,7 @@ static void PrintCatalogThread (CWnd* pParent, CFrameWnd* pMainFrame)
 //			dlgPrintStatus.SetDlgItemText(AFX_IDC_PRINT_DOCNAME, node->name);
 			node = node->next;
 			PieceInfo* pInfo = pLib->GetPieceInfo(node->actual);
-			pInfo->ZoomExtents(30.0f, (float)aspect);
+			pInfo->ZoomExtents(30.0f, aspect);
 
 			float pos[4] = { 0, 0, 10, 0 };
 			glLightfv(GL_LIGHT0, GL_POSITION, pos);
@@ -578,7 +578,7 @@ static void PrintPiecesThread(void* pv)
 	OpenGLSetPixelFormat(pMemDC->m_hDC, pixelformat, &pfd);
   HGLRC hmemrc = pfnwglCreateContext(pMemDC->GetSafeHdc());
 	pfnwglMakeCurrent(pMemDC->GetSafeHdc(), hmemrc);
-	double aspect = (float)picw/(float)h;
+	float aspect = (float)picw/(float)h;
 	glMatrixMode(GL_MODELVIEW);
 	glViewport(0, 0, picw, h);
 
@@ -692,7 +692,7 @@ static void PrintPiecesThread(void* pv)
 
 			PieceInfo* pInfo = pLib->GetPieceInfo(node->actual);
 			node = node->next;
-			pInfo->ZoomExtents(30.0f, (float)aspect);
+			pInfo->ZoomExtents(30.0f, aspect);
 
 			float pos[4] = { 0, 0, 10, 0 };
 			glLightfv(GL_LIGHT0, GL_POSITION, pos);
